Created Post-Processing
The Post-Processing Package by Unity Technologies was imported. A new Post-Processing Profile was created and linked to the Post-Processing Volume. Volumes size is 10,7,5. Post-Processing Layer Component added to main camera. Added both Bloom and Grain to the profile and set Intensity to 1.3 and Threshold to 0.6 to Bloom. Grain’s Intensity set to 1 and size set to 2. Created a new layer for the camera 8: PostProcessing and set the camera layer to 8

Project: Enshi City